feat: prev image 설정 로직 추가 (#405)
